use guint as elemstyle_color_t
It doesn't make sense to use 8 byte on 64 bit machines.

+++ src/josm_elemstyles.h
@@ -42,7 +42,7 @@ typedef enum {
 
 #define DEFAULT_DASH_LENGTH 0
 
-typedef gulong elemstyle_color_t;
+typedef guint elemstyle_color_t;
 
 typedef struct elemstyle_t elemstyle_t;
